Cleanup device_type_from_poolid() in DevicePool.
Change return type to uint16_t to match the device type.

Change "Memory" and "Processor" to "MemoryPool" and "ProcessorPool" so that it's more strict.

Replace strstr with STARTS_WITH() so that the check is more strict.


 Virt_DevicePool.c |   12 ++++++------
 Virt_DevicePool.h |   10 +++++-----
 2 files changed, 11 insertions(+), 11 deletions(-)

+++ src/Virt_DevicePool.h
@@ -56,11 +56,11 @@ char *pool_member_of(const CMPIBroker *broker,
                      const char *id);
 
 /**
- *  * Get the device type of a given pool from the pool's InstanceID
- *   *
- *    * @param id The InstanceID of the pool
- *     */
-int device_type_from_poolid(const char *id);
+ * Get the device type of a given pool from the pool's InstanceID
+ *
+ * @param id The InstanceID of the pool
+ */
+uint16_t device_type_from_poolid(const char *id);
 
 /**
  * Get all device pools on the system for the given connection
